feat: NHN KCP 가상계좌 완성
- PC/모바일 결제창에 vcnt_expire_term(입금 유효기간), disp_tax_yn 파라미터 추가
- vbank_expire_days 설정 항목 추가 (어드민 UI + 번역 + RegisterPgProviderListener)
- authCallback에서 virtual_account/vbank 둘 다 VCNT로 감지하도록 수정
- VbankNotifyRequest: good_mny 검증을 numeric으로 완화 (KCP 문자열 전송 대응)
- 주문완료 페이지(user_order_complete_receipt)에 가상계좌 입금 안내 섹션 추가

+++ src/Http/Requests/VbankNotifyRequest.php
@@ -35,7 +35,7 @@ class VbankNotifyRequest extends FormRequest
         return [
             'tno' => ['required', 'string'],
             'ordr_idxx' => ['required', 'string'],
-            'good_mny' => ['required', 'integer', 'min:1'],
+            'good_mny' => ['required', 'numeric', 'min:1'],
             'res_cd' => ['required', 'string'],
             'res_msg' => ['nullable', 'string'],
             'bank_name' => ['nullable', 'string'],
